Add Field Limiting to Product Query
Previously, the product query did not support field limiting, which meant that all fields were returned in the response. This commit introduces field limiting functionality, allowing clients to specify which fields they want to include in the response. If no fields are specified, the `__v` field is excluded by default. This enhancement improves performance and reduces the amount of data transferred.