Add bcrypt for password hashing in user model
This commit introduces the bcrypt library to the project, enabling secure password hashing within the user model. The userSchema now includes a pre-save middleware that hashes the user's password before saving it to the database. This enhancement improves the security of user data by ensuring that passwords are not stored in plain text.

Additionally, the package-lock.json file has been updated to reflect the new dependency and its associated transitive dependencies. This includes updates to various packages such as @mapbox/node-pre-gyp, agent-base, node-fetch, and others, which are required for the proper functioning of bcrypt.

The package.json file has also been modified to include the bcrypt dependency.