Add password change tracking and token invalidation
Track when a user's password is changed by adding a `passwordChangedAt` field to the user model. Implement logic in the authentication service to invalidate tokens issued before the password change. Update the user service to set the `passwordChangedAt` field when a user changes their password and issue a new JWT token.