Add Route and Service for Deleting Logged-in User
This commit introduces a new route and corresponding service method to handle the deletion (deactivation) of a logged-in user. The new route `/deleteMe` is added to the `userRoute.js` file, which calls the `deleteLoggedUser` service method. The `deleteLoggedUser` method in `userService.js` updates the user's `active` status to `false`, effectively deactivating the user account. The response returns a 204 status code with a message indicating the user has been deleted.