Add xss-clean middleware for XSS protection
This commit introduces the `xss-clean` middleware to the project, enhancing security by sanitizing user inputs to prevent Cross-Site Scripting (XSS) attacks. The middleware is added to the `server.js` file and the necessary dependencies (`xss-clean` and `xss-filters`) are included in the `package.json` and `package-lock.json` files. This update ensures that any potentially harmful scripts are neutralized before they can be executed, thereby safeguarding the application from XSS vulnerabilities.